Add Users class and inheritance for Reader and Employee classes
